requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==4.2
django-redis==5.4.0
django_injector==0.3.1
djangorestframework==3.14.0
fastapi==0.115.6
idna==3.10
iniconfig==2.0.0
injector==0.22.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
pytest==7.4.2
pytest-django==4.5.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
requests==2.31.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.3.0
